rename parsing.py back to calc.py; move factorial error
