+++ tests/helpers/math/test_calc_arrays.py
